requirements.txt
discord.py
plex.py==0.9.0
PlexAPI==4.13.0
texttable
python-dotenv
